Gtyw embed (#670)
* gtwy embed

* add condition for emebed

* add modal using global modal

---------

+++ utils/interceptor.js
@@ -2,7 +2,7 @@ import axios from "axios";
 
 axios.interceptors.request.use(
     async (config) => {
-        let token = localStorage.getItem("proxy_token");
+        let token = sessionStorage.getItem("proxy_token") ? sessionStorage.getItem("proxy_token") : localStorage.getItem("proxy_token");
         config.headers['proxy_auth_token'] = token;
         if (process.env.NEXT_PUBLIC_ENV === 'local')
             config.headers['Authorization'] = localStorage.getItem("local_token");
@@ -19,9 +19,9 @@ axios.interceptors.response.use(
         return response;
     },
     async function (error) {
-        if (error?.response?.status === 401) {
+        if (error?.response?.status === 401 && !sessionStorage.getItem("proxy_token")) {
             localStorage.clear();
-            if(window.location.href!='/login')localStorage.setItem("previous_url", window.location.href);           
+            if (window.location.href != '/login') localStorage.setItem("previous_url", window.location.href);
             window.location.href = "/login";
         }
         return Promise.reject(error);
